+++ Assets/Scripts/RidersGame/Types.cs
@@ -30,4 +30,16 @@ namespace RidersRuntime.Data
     {
         public const string DebugFlags = "debug_flags.json";
     }
+
+    public enum GameScene
+    {
+        TitleScreen = 0,
+        MainMenu = 1,
+        Loading = 2,
+        Game = 3,
+        TutorialArea = 4,
+        FirstRaceInMelbourne = 5
+        // Add more scenes as needed
+    }
+
 }
