--- Assets/Scripts/UI/MenuScreen/MenuUI.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public class MenuUI : MonoBehaviour
-{
-    public void LoadMainMenu()
-    {
-        SceneLoader.LoadScene(Scene.MainMenu);
-    }
-}
