--- Assets/Scripts/UI/SceneManagement/SceneEnum.cs
@@ -1,8 +1,0 @@
-public enum Scene
-{
-    TitleScreen = 0,
-    MainMenu = 1,
-    Loading = 2,
-    Game = 3
-    // Add more scenes as needed
-}
