--- Assets/Scripts/UI/SceneManagement/SceneEnum.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 375c3a2dab90f4bc0a06030776039438
